requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
coverage==6.3.1
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.0
Markdown==3.3.6
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
